requirements.txt
streamlit
numpy
Pillow
torch
torchvision
scikit-learn
git+https://github.com/openai/CLIP.git